Use alias to override yes-or-no-p for easier reversibility
`fset' binds a definition to a symbol, which becomes difficult to
reverse if the user would prefer the longer 'yes'/'no' functionality
over the 'y'/'n' functionality. Using `fmakunbind' makes the function
void and using another `fset' causes either an infinite recursion,
which locks emacs or a call to a void function or possibly a call with
improper arguments depending on the use and/or timing of
`fmakunbind'.

Using an alias is easy to revert simply by calling `advice-remove'.