Create rational-emacs module for initialization.
Attempts to address issue #59.
Fix for issue #62.

The contents of early-init.el and init.el have moved to functions
in the rational-emacs.el file.

early-init.el and init.el are updated to call the appropriate
functions in this new module.

All functions are updated to reflect the module name 'rational-emacs'.