Use package.el to install use-package
It seemed odd to use straight to install `use-package', when the rest
of this module is about using package.el. Update priorities to pull
from gnu/nongnu elpas over melpa per the principles, install
`use-package' from one of the elpas if it is not already
installed. Also do not activate packages when calling
`package-initialize', prefer to allow `use-package', `require', or
`autoload' to activate packages as needed by the user.