Fix rational modules skeleton regex
Expected the regex to only apply to the filename without the path,
however, the code uses `buffer-file-name` which includes the path. So
personal modules whose path matched the regex resulted in the module
skeleton insertion which was not intended. This should fix #89 and
only insert for elisp files found in the modules folder and named
rational-<module>.el