Correct the Example Configuration; Add more commentary to rational-ui.el
The example configuration now uses an unquoted symbol in the weight face attribute plist, as M-x `customize-group` "rational" would do when customizing the variable.

Using the example configuration on the current HEAD causes an error, because the symbolp fails; this may confuse some users, because the predicate will pass interactively.