Introduce "Sensible Folder Structure"
This aims to setup initial infrastructure to keep state files
inside `crafter-config-var-directory`.

- A setting is created to enable/disable the behaviour (enabled by default)
- A helper function is introduced to ease adding new state files
- This is all documented

Unrelated to the above is:

We add `CUSTOM_ID` to a documentation entry to enable creating an internal link.

This is because according to [orgmode documentation](https://orgmode.org/manual/Internal-Links.html) local
links by ID work only when `CUSTOM_ID` is set.

> (...) a construct like ‘[[#my-custom-id]]’  specifically targets the
entry with the ‘CUSTOM_ID’ property set to ‘my-custom-id’.

Attempting to export `README.org` without this results in an error.