Fix rational-compile for issue #153
Still have a strange bug with `native-compile`.

If you start your emacs session configured with `rational-compile` to
compile modules, init or config on startup, it will stall on a second
startup, complaining that `rational-package-install-package` is not
available.

The issue seems to be solved once you delete all `.elc` files from
`user-emacs-directory` and `rational-config-path`, and only then
`rational-compile-init`, `rational-compile-config` or
`rational-compile-modules`.

I don't know if that problem is specific to my system/configuration,
but I have tried daviwil's configuration with only `rational-compile`
added to it, and it crashes.

Maybe if `rational-package-install-package` is turned into a module,
and that module is required on each module that uses it, it could stop
complaining that it is not available.

Please read, comment, make suggestions and corrections.