Resolve issue showing startup screen in emacsclient
Using the `after-init-hook` works, but when running emacs as a daemon,
the hook is run when the server is started and the startup screen is
not shown in clients. This is resolved by using the
`initial-buffer-choice` variable and setting it to show the startup
screen unless the user has set the rational-startup-inhibit-splash to
a non-nil value.