Update comments in bootstrap code
The intention is to use these in crafted modules as a consistent
interface to the package system.  Since we can't predict which system
the user will ultimately use (currently only straight and package are
supported), we need a consistent interface to "do the right thing".
The user is not expected to use this code, they may if they choose, or
they may choose a different front-end (ala use-package or leaf, etc).