Remove doom-modeline due to upstream instability.
The doom-modeline package has caused a lot of grief for users not
following the development branch, often breaking their
configurations.  While there are solutions, the number of recent
issues and frustration in the community outweighs the value of keeping
and managing the things that break when changes are made to this
package.