Remove customization for use-short-answers
The docstring for use-short-answers does not recommend setting it to a
non-nil value as the intention is to have the user think about what
they are doing rather than answering too quickly.  Crafted Emacs will
support this position, the user can customize this value in their own
configuration if they choose.