Switch to customize-save-variable in certain cases
Because we defer customization of some packages until after they are
loaded, and those packages may not be loaded until one of their
autoload functions is run, the customizations would not be save after
emacs is finished initializing.  As we prefer to save those
customizations, switch to using 'customize-save-variable' instead to
make sure it is set and saved by the customization system for next
time.