Fix #276 an be more consistent when checking Emacs versions
Per feedback provided in PR #279, updated, where necessary, checks for
Emacs versions.  The version< function uses strings as parameters, but
is purpose written to handle Emacs version comparisons.  Now provides
consistent approach to checking the Emacs version for specific
functionality.  Thanks to @japhir for the original PR and attention to
this issue.