Update crafted-ide to make tree-sitter optional
Introduces three new functions, one of which is user-facing.

`crafted-ide--configure-tree-sitter-pre-29` sets up tree-sitter pre-Emacs 29.

`crafted-ide--configure-tree-sitter` sets up tree-sitter for Emacs 29 or later.
An list of language grammar symbols can be passed to opt-out of certain
language grammars.

`crafted-ide-configure-tree-sitter` automatically selects one of the functions
above, depending on the detected version of Emacs.
The opt-out list parameter is optional.
For Emacs 29 or later, the opt-out list is passed through to the function.