[v2] Fixes #375, remove combobulate from package-selected-packages
We note in the documentation combobulate is not available on the
normal package repos (M)ELPA etc, so we should not add it to the list
of package-selected-packages.

The user can install it via an alternative package manager
(package-vc, straight, borg, etc): or manually following the
instructions from the GitHub repo (which is called out in our docs).