Attempt fix for #390
Stop using `project' related functions to find the project root as
this seems to be not working in cases where the source is downloaded
or when the project is added as a submodule.  This also simplifies the
code a bit.

Add throwing an error if the `crafted-emacs-home' variable is still
not set after trying to automatically figure it out, but before
updateing the `load-path'.  In that case, it is impossible to update
the `load-path' appropriately and trying to load crafted-emacs modules
will fail, so throw an error with a message to indicate what needs to
be done.