Change symantics for treesit-auto configuration to opt-in
Versions of treesit-auto after 0.6.6 remove the variable
treesit-auto-opt-out-list which we were using to opt out of certain
languages that caused problems, for example proto/protobuf.

treesit-auto always seemed to provide a way to limit the installed ts
modes to just the list in treesit-auto-langs, thus the user should
customize this variable appropriately, or accept the default of all
language ts modes.

crafted-ide--configure-tree-sitter changed to use the optional
provided list as the "opt-in" list rather than the "opt out" list.

Fixes #402